saasherder TargetSpec handle parameters
